Extract do_turn from game
