+++ src/do_turn.cpp
@@ -1,0 +1,1050 @@
+#include "do_turn.h"
+
+#include "action.h"
+#include "avatar.h"
+#include "bionics.h"
+#include "calendar.h"
+#include "event_bus.h"
+#include "explosion.h"
+#include "game.h"
+#include "gamemode.h"
+#include "help.h"
+#include "kill_tracker.h"
+#include "make_static.h"
+#include "map.h"
+#include "mapbuffer.h"
+#include "memorial_logger.h"
+#include "messages.h"
+#include "mission.h"
+#include "monattack.h"
+#include "mtype.h"
+#include "options.h"
+#include "output.h"
+#include "overmapbuffer.h"
+#include "popup.h"
+#include "scent_map.h"
+#include "string_input_popup.h"
+#include "timed_event.h"
+#include "ui_manager.h"
+#include "vehicle.h"
+#include "vpart_position.h"
+#include "wcwidth.h"
+#include "worldfactory.h"
+
+static const activity_id ACT_OPERATION( "ACT_OPERATION" );
+
+static const efftype_id effect_controlled( "controlled" );
+static const efftype_id effect_downed( "downed" );
+static const efftype_id effect_npc_suspend( "npc_suspend" );
+static const efftype_id effect_ridden( "ridden" );
+static const efftype_id effect_sleep( "sleep" );
+
+static const itype_id itype_holybook_bible1( "holybook_bible1" );
+static const itype_id itype_holybook_bible2( "holybook_bible2" );
+static const itype_id itype_holybook_bible3( "holybook_bible3" );
+
+static const trait_id trait_LEG_TENT_BRACE( "LEG_TENT_BRACE" );
+
+#define dbg(x) DebugLog((x),D_GAME) << __FILE__ << ":" << __LINE__ << ": "
+
+namespace turn_handler
+{
+bool cleanup_at_end()
+{
+    if( g->uquit == QUIT_DIED || g->uquit == QUIT_SUICIDE ) {
+        // Put (non-hallucinations) into the overmap so they are not lost.
+        for( monster &critter : g->all_monsters() ) {
+            g->despawn_monster( critter );
+        }
+        // Reset NPC factions and disposition
+        g->reset_npc_dispositions();
+        // Save the factions', missions and set the NPC's overmap coordinates
+        // Npcs are saved in the overmap.
+        g->save_factions_missions_npcs(); //missions need to be saved as they are global for all saves.
+
+        // and the overmap, and the local map.
+        g->save_maps(); //Omap also contains the npcs who need to be saved.
+    }
+
+    avatar &u = get_avatar();
+    if( g->uquit == QUIT_DIED || g->uquit == QUIT_SUICIDE ) {
+        std::vector<std::string> vRip;
+
+        int iMaxWidth = 0;
+        int iNameLine = 0;
+        int iInfoLine = 0;
+
+        if( u.has_amount( itype_holybook_bible1, 1 ) || u.has_amount( itype_holybook_bible2, 1 ) ||
+            u.has_amount( itype_holybook_bible3, 1 ) ) {
+            if( !( u.has_trait( trait_id( "CANNIBAL" ) ) || u.has_trait( trait_id( "PSYCHOPATH" ) ) ) ) {
+                vRip.emplace_back( "               _______  ___" );
+                vRip.emplace_back( "              <       `/   |" );
+                vRip.emplace_back( "               >  _     _ (" );
+                vRip.emplace_back( "              |  |_) | |_) |" );
+                vRip.emplace_back( "              |  | \\ | |   |" );
+                vRip.emplace_back( "   ______.__%_|            |_________  __" );
+                vRip.emplace_back( " _/                                  \\|  |" );
+                iNameLine = vRip.size();
+                vRip.emplace_back( "|                                        <" );
+                vRip.emplace_back( "|                                        |" );
+                iMaxWidth = utf8_width( vRip.back() );
+                vRip.emplace_back( "|                                        |" );
+                vRip.emplace_back( "|_____.-._____              __/|_________|" );
+                vRip.emplace_back( "              |            |" );
+                iInfoLine = vRip.size();
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |           <" );
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |   _        |" );
+                vRip.emplace_back( "              |__/         |" );
+                vRip.emplace_back( "             % / `--.      |%" );
+                vRip.emplace_back( "         * .%%|          -< @%%%" ); // NOLINT(cata-text-style)
+                vRip.emplace_back( "         `\\%`@|            |@@%@%%" );
+                vRip.emplace_back( "       .%%%@@@|%     `   % @@@%%@%%%%" );
+                vRip.emplace_back( "  _.%%%%%%@@@@@@%%%__/\\%@@%%@@@@@@@%%%%%%" );
+
+            } else {
+                vRip.emplace_back( "               _______  ___" );
+                vRip.emplace_back( "              |       \\/   |" );
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |            |" );
+                iInfoLine = vRip.size();
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |            |" );
+                vRip.emplace_back( "              |           <" );
+                vRip.emplace_back( "              |   _        |" );
+                vRip.emplace_back( "              |__/         |" );
+                vRip.emplace_back( "   ______.__%_|            |__________  _" );
+                vRip.emplace_back( " _/                                   \\| \\" );
+                iNameLine = vRip.size();
+                vRip.emplace_back( "|                                         <" );
+                vRip.emplace_back( "|                                         |" );
+                iMaxWidth = utf8_width( vRip.back() );
+                vRip.emplace_back( "|                                         |" );
+                vRip.emplace_back( "|_____.-._______            __/|__________|" );
+                vRip.emplace_back( "             % / `_-.   _  |%" );
+                vRip.emplace_back( "         * .%%|  |_) | |_)< @%%%" ); // NOLINT(cata-text-style)
+                vRip.emplace_back( "         `\\%`@|  | \\ | |   |@@%@%%" );
+                vRip.emplace_back( "       .%%%@@@|%     `   % @@@%%@%%%%" );
+                vRip.emplace_back( "  _.%%%%%%@@@@@@%%%__/\\%@@%%@@@@@@@%%%%%%" );
+            }
+        } else {
+            vRip.emplace_back( R"(           _________  ____           )" );
+            vRip.emplace_back( R"(         _/         `/    \_         )" );
+            vRip.emplace_back( R"(       _/      _     _      \_.      )" );
+            vRip.emplace_back( R"(     _%\      |_) | |_)       \_     )" );
+            vRip.emplace_back( R"(   _/ \/      | \ | |           \_   )" );
+            vRip.emplace_back( R"( _/                               \_ )" );
+            vRip.emplace_back( R"(|                                   |)" );
+            iNameLine = vRip.size();
+            vRip.emplace_back( R"( )                                 < )" );
+            vRip.emplace_back( R"(|                                   |)" );
+            vRip.emplace_back( R"(|                                   |)" );
+            vRip.emplace_back( R"(|   _                               |)" );
+            vRip.emplace_back( R"(|__/                                |)" );
+            iMaxWidth = utf8_width( vRip.back() );
+            vRip.emplace_back( R"( / `--.                             |)" );
+            vRip.emplace_back( R"(|                                  ( )" );
+            iInfoLine = vRip.size();
+            vRip.emplace_back( R"(|                                   |)" );
+            vRip.emplace_back( R"(|                                   |)" );
+            vRip.emplace_back( R"(|     %                         .   |)" );
+            vRip.emplace_back( R"(|  @`                            %% |)" );
+            vRip.emplace_back( R"(| %@%@%\                *      %`%@%|)" );
+            vRip.emplace_back( R"(%%@@@.%@%\%%            `\  %%.%%@@%@)" );
+            vRip.emplace_back( R"(@%@@%%%%%@@@@@@%%%%%%%%@@%%@@@%%%@%%@)" );
+        }
+
+        const point iOffset( TERMX > FULL_SCREEN_WIDTH ? ( TERMX - FULL_SCREEN_WIDTH ) / 2 : 0,
+                             TERMY > FULL_SCREEN_HEIGHT ? ( TERMY - FULL_SCREEN_HEIGHT ) / 2 : 0 );
+
+        catacurses::window w_rip = catacurses::newwin( FULL_SCREEN_HEIGHT, FULL_SCREEN_WIDTH,
+                                   iOffset );
+        draw_border( w_rip );
+
+        sfx::do_player_death_hurt( get_player_character(), true );
+        sfx::fade_audio_group( sfx::group::weather, 2000 );
+        sfx::fade_audio_group( sfx::group::time_of_day, 2000 );
+        sfx::fade_audio_group( sfx::group::context_themes, 2000 );
+        sfx::fade_audio_group( sfx::group::fatigue, 2000 );
+
+        for( size_t iY = 0; iY < vRip.size(); ++iY ) {
+            size_t iX = 0;
+            const char *str = vRip[iY].data();
+            for( int slen = vRip[iY].size(); slen > 0; ) {
+                const uint32_t cTemp = UTF8_getch( &str, &slen );
+                if( cTemp != U' ' ) {
+                    nc_color ncColor = c_light_gray;
+
+                    if( cTemp == U'%' ) {
+                        ncColor = c_green;
+
+                    } else if( cTemp == U'_' || cTemp == U'|' ) {
+                        ncColor = c_white;
+
+                    } else if( cTemp == U'@' ) {
+                        ncColor = c_brown;
+
+                    } else if( cTemp == U'*' ) {
+                        ncColor = c_red;
+                    }
+
+                    mvwputch( w_rip, point( iX + FULL_SCREEN_WIDTH / 2 - ( iMaxWidth / 2 ), iY + 1 ), ncColor,
+                              cTemp );
+                }
+                iX += mk_wcwidth( cTemp );
+            }
+        }
+
+        std::string sTemp;
+
+        center_print( w_rip, iInfoLine++, c_white, _( "Survived:" ) );
+
+        const time_duration survived = calendar::turn - calendar::start_of_game;
+        const int minutes = to_minutes<int>( survived ) % 60;
+        const int hours = to_hours<int>( survived ) % 24;
+        const int days = to_days<int>( survived );
+
+        if( days > 0 ) {
+            sTemp = string_format( "%dd %dh %dm", days, hours, minutes );
+        } else if( hours > 0 ) {
+            sTemp = string_format( "%dh %dm", hours, minutes );
+        } else {
+            sTemp = string_format( "%dm", minutes );
+        }
+
+        center_print( w_rip, iInfoLine++, c_white, sTemp );
+
+        const int iTotalKills = g->get_kill_tracker().monster_kill_count();
+
+        sTemp = _( "Kills:" );
+        mvwprintz( w_rip, point( FULL_SCREEN_WIDTH / 2 - 5, 1 + iInfoLine++ ), c_light_gray,
+                   ( sTemp + " " ) );
+        wprintz( w_rip, c_magenta, "%d", iTotalKills );
+
+        sTemp = _( "In memory of:" );
+        mvwprintz( w_rip, point( FULL_SCREEN_WIDTH / 2 - utf8_width( sTemp ) / 2, iNameLine++ ),
+                   c_light_gray,
+                   sTemp );
+
+        sTemp = u.name;
+        mvwprintz( w_rip, point( FULL_SCREEN_WIDTH / 2 - utf8_width( sTemp ) / 2, iNameLine++ ), c_white,
+                   sTemp );
+
+        sTemp = _( "Last Words:" );
+        mvwprintz( w_rip, point( FULL_SCREEN_WIDTH / 2 - utf8_width( sTemp ) / 2, iNameLine++ ),
+                   c_light_gray,
+                   sTemp );
+
+        int iStartX = FULL_SCREEN_WIDTH / 2 - ( ( iMaxWidth - 4 ) / 2 );
+        std::string sLastWords = string_input_popup()
+                                 .window( w_rip, point( iStartX, iNameLine ), iStartX + iMaxWidth - 4 - 1 )
+                                 .max_length( iMaxWidth - 4 - 1 )
+                                 .query_string();
+        g->death_screen();
+        const bool is_suicide = g->uquit == QUIT_SUICIDE;
+        std::chrono::seconds time_since_load =
+            std::chrono::duration_cast<std::chrono::seconds>(
+                std::chrono::steady_clock::now() - g->time_of_last_load );
+        std::chrono::seconds total_time_played = g->time_played_at_last_load + time_since_load;
+        get_event_bus().send<event_type::game_over>( is_suicide, sLastWords, total_time_played );
+        // Struck the save_player_data here to forestall Weirdness
+        g->move_save_to_graveyard();
+        g->write_memorial_file( sLastWords );
+        get_memorial().clear();
+        std::vector<std::string> characters = g->list_active_characters();
+        // remove current player from the active characters list, as they are dead
+        std::vector<std::string>::iterator curchar = std::find( characters.begin(),
+                characters.end(), u.name );
+        if( curchar != characters.end() ) {
+            characters.erase( curchar );
+        }
+
+        if( characters.empty() ) {
+            bool queryDelete = false;
+            bool queryReset = false;
+
+            if( get_option<std::string>( "WORLD_END" ) == "query" ) {
+                bool decided = false;
+                std::string buffer = _( "Warning: NPC interactions and some other global flags "
+                                        "will not all reset when starting a new character in an "
+                                        "already-played world.  This can lead to some strange "
+                                        "behavior.\n\n"
+                                        "Are you sure you wish to keep this world?"
+                                      );
+
+                while( !decided ) {
+                    uilist smenu;
+                    smenu.allow_cancel = false;
+                    smenu.addentry( 0, true, 'r', "%s", _( "Reset world" ) );
+                    smenu.addentry( 1, true, 'd', "%s", _( "Delete world" ) );
+                    smenu.addentry( 2, true, 'k', "%s", _( "Keep world" ) );
+                    smenu.query();
+
+                    switch( smenu.ret ) {
+                        case 0:
+                            queryReset = true;
+                            decided = true;
+                            break;
+                        case 1:
+                            queryDelete = true;
+                            decided = true;
+                            break;
+                        case 2:
+                            decided = query_yn( buffer );
+                            break;
+                    }
+                }
+            }
+
+            if( queryDelete || get_option<std::string>( "WORLD_END" ) == "delete" ) {
+                world_generator->delete_world( world_generator->active_world->world_name, true );
+
+            } else if( queryReset || get_option<std::string>( "WORLD_END" ) == "reset" ) {
+                world_generator->delete_world( world_generator->active_world->world_name, false );
+            }
+        } else if( get_option<std::string>( "WORLD_END" ) != "keep" ) {
+            std::string tmpmessage;
+            for( auto &character : characters ) {
+                tmpmessage += "\n  ";
+                tmpmessage += character;
+            }
+            popup( _( "World retained.  Characters remaining:%s" ), tmpmessage );
+        }
+        if( g->gamemode ) {
+            g->gamemode = std::make_unique<special_game>(); // null gamemode or something..
+        }
+    }
+
+    //Reset any offset due to driving
+    g->set_driving_view_offset( point_zero );
+
+    //clear all sound channels
+    sfx::fade_audio_channel( sfx::channel::any, 300 );
+    sfx::fade_audio_group( sfx::group::weather, 300 );
+    sfx::fade_audio_group( sfx::group::time_of_day, 300 );
+    sfx::fade_audio_group( sfx::group::context_themes, 300 );
+    sfx::fade_audio_group( sfx::group::fatigue, 300 );
+
+    zone_manager::get_manager().clear();
+
+    MAPBUFFER.clear();
+    overmap_buffer.clear();
+
+#if defined(__ANDROID__)
+    quick_shortcuts_map.clear();
+#endif
+    return true;
+}
+
+// TODO: abstract out the location checking code
+// TODO: refactor so zombies can follow up and down stairs instead of this mess
+void update_stair_monsters()
+{
+    // Search for the stairs closest to the player.
+    std::vector<int> stairx;
+    std::vector<int> stairy;
+    std::vector<int> stairdist;
+
+    map &m = get_map();
+    const bool from_below = g->monstairz < m.get_abs_sub().z;
+
+    if( g->coming_to_stairs.empty() ) {
+        return;
+    }
+
+    if( m.has_zlevels() ) {
+        debugmsg( "%d monsters coming to stairs on a map with z-levels",
+                  g->coming_to_stairs.size() );
+        g->coming_to_stairs.clear();
+    }
+
+    avatar &u = get_avatar();
+    for( const tripoint &dest : m.points_on_zlevel( u.posz() ) ) {
+        if( ( from_below && m.has_flag( "GOES_DOWN", dest ) ) ||
+            ( !from_below && m.has_flag( "GOES_UP", dest ) ) ) {
+            stairx.push_back( dest.x );
+            stairy.push_back( dest.y );
+            stairdist.push_back( rl_dist( dest, u.pos() ) );
+        }
+    }
+    if( stairdist.empty() ) {
+        return;         // Found no stairs?
+    }
+
+    // Find closest stairs.
+    size_t si = 0;
+    for( size_t i = 0; i < stairdist.size(); i++ ) {
+        if( stairdist[i] < stairdist[si] ) {
+            si = i;
+        }
+    }
+
+    // Find up to 4 stairs for distance stairdist[si] +1
+    std::vector<int> nearest;
+    nearest.push_back( si );
+    for( size_t i = 0; i < stairdist.size() && nearest.size() < 4; i++ ) {
+        if( ( i != si ) && ( stairdist[i] <= stairdist[si] + 1 ) ) {
+            nearest.push_back( i );
+        }
+    }
+    // Randomize the stair choice
+    si = random_entry_ref( nearest );
+
+    // Attempt to spawn zombies.
+    for( size_t i = 0; i < g->coming_to_stairs.size(); i++ ) {
+        point mpos( stairx[si], stairy[si] );
+        monster &critter = g->coming_to_stairs[i];
+        const tripoint dest {
+            mpos, get_map().get_abs_sub().z
+        };
+
+        // We might be not be visible.
+        if( ( critter.posx() < 0 - ( MAPSIZE_X ) / 6 ||
+              critter.posy() < 0 - ( MAPSIZE_Y ) / 6 ||
+              critter.posx() > ( MAPSIZE_X * 7 ) / 6 ||
+              critter.posy() > ( MAPSIZE_Y * 7 ) / 6 ) ) {
+            continue;
+        }
+
+        critter.staircount -= 4;
+        // Let the player know zombies are trying to come.
+        if( u.sees( dest ) ) {
+            std::string dump;
+            if( critter.staircount > 4 ) {
+                dump += string_format( _( "You see a %s on the stairs" ), critter.name() );
+            } else {
+                if( critter.staircount > 0 ) {
+                    dump += ( from_below ?
+                              //~ The <monster> is almost at the <bottom/top> of the <terrain type>!
+                              string_format( _( "The %1$s is almost at the top of the %2$s!" ),
+                                             critter.name(),
+                                             m.tername( dest ) ) :
+                              string_format( _( "The %1$s is almost at the bottom of the %2$s!" ),
+                                             critter.name(),
+                                             m.tername( dest ) ) );
+                }
+            }
+
+            add_msg( m_warning, dump );
+        } else {
+            sounds::sound( dest, 5, sounds::sound_t::movement,
+                           _( "a sound nearby from the stairs!" ), true, "misc", "stairs_movement" );
+        }
+
+        if( critter.staircount > 0 ) {
+            continue;
+        }
+
+        if( g->is_empty( dest ) ) {
+            critter.spawn( dest );
+            critter.staircount = 0;
+            g->place_critter_at( make_shared_fast<monster>( critter ), dest );
+            if( u.sees( dest ) ) {
+                if( !from_below ) {
+                    add_msg( m_warning, _( "The %1$s comes down the %2$s!" ),
+                             critter.name(),
+                             m.tername( dest ) );
+                } else {
+                    add_msg( m_warning, _( "The %1$s comes up the %2$s!" ),
+                             critter.name(),
+                             m.tername( dest ) );
+                }
+            }
+            g->coming_to_stairs.erase( g->coming_to_stairs.begin() + i );
+            continue;
+        } else if( u.pos() == dest ) {
+            // Monster attempts to push player of stairs
+            point push( point_north_west );
+            int tries = 0;
+
+            // the critter is now right on top of you and will attack unless
+            // it can find a square to push you into with one of his tries.
+            const int creature_push_attempts = 9;
+            const int player_throw_resist_chance = 3;
+
+            critter.spawn( dest );
+            while( tries < creature_push_attempts ) {
+                tries++;
+                push.x = rng( -1, 1 );
+                push.y = rng( -1, 1 );
+                point ipos( mpos + push );
+                tripoint pos( ipos, m.get_abs_sub().z );
+                if( ( push.x != 0 || push.y != 0 ) && !g->critter_at( pos ) &&
+                    critter.can_move_to( pos ) ) {
+                    bool resiststhrow = ( u.is_throw_immune() ) ||
+                                        ( u.has_trait( trait_LEG_TENT_BRACE ) );
+                    if( resiststhrow && one_in( player_throw_resist_chance ) ) {
+                        u.moves -= 25; // small charge for avoiding the push altogether
+                        add_msg( _( "The %s fails to push you back!" ),
+                                 critter.name() );
+                        return; //judo or leg brace prevent you from getting pushed at all
+                    }
+                    // Not accounting for tentacles latching on, so..
+                    // Something is about to happen, lets charge half a move
+                    u.moves -= 50;
+                    if( resiststhrow && ( u.is_throw_immune() ) ) {
+                        //we have a judoka who isn't getting pushed but counterattacking now.
+                        mattack::thrown_by_judo( &critter );
+                        return;
+                    }
+                    std::string msg;
+                    ///\EFFECT_DODGE reduces chance of being downed when pushed off the stairs
+                    if( !( resiststhrow ) && ( u.get_dodge() + rng( 0, 3 ) < 12 ) ) {
+                        // dodge 12 - never get downed
+                        // 11.. avoid 75%; 10.. avoid 50%; 9.. avoid 25%
+                        u.add_effect( effect_downed, 2_turns );
+                        msg = _( "The %s pushed you back hard!" );
+                    } else {
+                        msg = _( "The %s pushed you back!" );
+                    }
+                    add_msg( m_warning, msg.c_str(), critter.name() );
+                    u.setx( u.posx() + push.x );
+                    u.sety( u.posy() + push.y );
+                    return;
+                }
+            }
+            add_msg( m_warning,
+                     _( "The %s tried to push you back but failed!  It attacks you!" ),
+                     critter.name() );
+            critter.melee_attack( u );
+            u.moves -= 50;
+            return;
+        } else if( monster *const mon_ptr = g->critter_at<monster>( dest ) ) {
+            // Monster attempts to displace a monster from the stairs
+            monster &other = *mon_ptr;
+            critter.spawn( dest );
+
+            // the critter is now right on top of another and will push it
+            // if it can find a square to push it into inside of his tries.
+            const int creature_push_attempts = 9;
+            const int creature_throw_resist = 4;
+
+            int tries = 0;
+            point push2;
+            while( tries < creature_push_attempts ) {
+                tries++;
+                push2.x = rng( -1, 1 );
+                push2.y = rng( -1, 1 );
+                point ipos2( mpos + push2 );
+                tripoint pos( ipos2, m.get_abs_sub().z );
+                if( ( push2.x == 0 && push2.y == 0 ) || ( ( ipos2.x == u.posx() ) && ( ipos2.y == u.posy() ) ) ) {
+                    continue;
+                }
+                if( !g->critter_at( pos ) && other.can_move_to( pos ) ) {
+                    other.setpos( tripoint( ipos2, m.get_abs_sub().z ) );
+                    other.moves -= 50;
+                    std::string msg;
+                    if( one_in( creature_throw_resist ) ) {
+                        other.add_effect( effect_downed, 2_turns );
+                        msg = _( "The %1$s pushed the %2$s hard." );
+                    } else {
+                        msg = _( "The %1$s pushed the %2$s." );
+                    }
+                    add_msg( m_neutral, msg, critter.name(), other.name() );
+                    return;
+                }
+            }
+            return;
+        }
+    }
+}
+
+} // namespace turn_handler
+
+namespace
+{
+void handle_key_blocking_activity();
+void monmove();
+void overmap_npc_move();
+
+void handle_key_blocking_activity()
+{
+    avatar &u = get_avatar();
+    if( ( u.activity && u.activity.moves_left > 0 ) ||
+        ( u.has_destination() && !u.omt_path.empty() ) ) {
+        input_context ctxt = get_default_mode_input_context();
+        const std::string action = ctxt.handle_input( 0 );
+        bool refresh = true;
+        if( action == "pause" ) {
+            if( u.activity.interruptable_with_kb ) {
+                g->cancel_activity_query( _( "Confirm:" ) );
+            }
+        } else if( action == "player_data" ) {
+            u.disp_info();
+        } else if( action == "messages" ) {
+            Messages::display_messages();
+        } else if( action == "help" ) {
+            get_help().display_help();
+        } else if( action != "HELP_KEYBINDINGS" ) {
+            refresh = false;
+        }
+        if( refresh ) {
+            ui_manager::redraw();
+            refresh_display();
+        }
+    }
+}
+
+void monmove()
+{
+    g->cleanup_dead();
+    map &m = get_map();
+    avatar &u = get_avatar();
+
+    for( monster &critter : g->all_monsters() ) {
+        // Critters in impassable tiles get pushed away, unless it's not impassable for them
+        if( !critter.is_dead() && m.impassable( critter.pos() ) && !critter.can_move_to( critter.pos() ) ) {
+            dbg( D_ERROR ) << "game:monmove: " << critter.name()
+                           << " can't move to its location!  (" << critter.posx()
+                           << ":" << critter.posy() << ":" << critter.posz() << "), "
+                           << m.tername( critter.pos() );
+            add_msg_debug( debugmode::DF_MONSTER, "%s can't move to its location!  (%d,%d,%d), %s",
+                           critter.name(),
+                           critter.posx(), critter.posy(), critter.posz(), m.tername( critter.pos() ) );
+            bool okay = false;
+            for( const tripoint &dest : m.points_in_radius( critter.pos(), 3 ) ) {
+                if( critter.can_move_to( dest ) && g->is_empty( dest ) ) {
+                    critter.setpos( dest );
+                    okay = true;
+                    break;
+                }
+            }
+            if( !okay ) {
+                // die of "natural" cause (overpopulation is natural)
+                critter.die( nullptr );
+            }
+        }
+
+        if( !critter.is_dead() ) {
+            critter.process_turn();
+        }
+
+        m.creature_in_field( critter );
+        if( calendar::once_every( 1_days ) ) {
+            if( critter.has_flag( MF_MILKABLE ) ) {
+                critter.refill_udders();
+            }
+            critter.try_biosignature();
+            critter.try_reproduce();
+        }
+        while( critter.moves > 0 && !critter.is_dead() && !critter.has_effect( effect_ridden ) ) {
+            critter.made_footstep = false;
+            // Controlled critters don't make their own plans
+            if( !critter.has_effect( effect_controlled ) ) {
+                // Formulate a path to follow
+                critter.plan();
+            } else {
+                critter.moves = 0;
+                break;
+            }
+            critter.move(); // Move one square, possibly hit u
+            critter.process_triggers();
+            m.creature_in_field( critter );
+        }
+
+        const bionic_id bio_alarm( "bio_alarm" );
+        if( !critter.is_dead() &&
+            u.has_active_bionic( bio_alarm ) &&
+            u.get_power_level() >= bio_alarm->power_trigger &&
+            rl_dist( u.pos(), critter.pos() ) <= 5 &&
+            !critter.is_hallucination() ) {
+            u.mod_power_level( -bio_alarm->power_trigger );
+            add_msg( m_warning, _( "Your motion alarm goes off!" ) );
+            g->cancel_activity_or_ignore_query( distraction_type::motion_alarm,
+                                                _( "Your motion alarm goes off!" ) );
+            if( u.has_effect( efftype_id( "sleep" ) ) ) {
+                u.wake_up();
+            }
+        }
+    }
+
+    g->cleanup_dead();
+
+    // The remaining monsters are all alive, but may be outside of the reality bubble.
+    // If so, despawn them. This is not the same as dying, they will be stored for later and the
+    // monster::die function is not called.
+    for( monster &critter : g->all_monsters() ) {
+        if( critter.posx() < 0 - ( MAPSIZE_X ) / 6 ||
+            critter.posy() < 0 - ( MAPSIZE_Y ) / 6 ||
+            critter.posx() > ( MAPSIZE_X * 7 ) / 6 ||
+            critter.posy() > ( MAPSIZE_Y * 7 ) / 6 ) {
+            g->despawn_monster( critter );
+        }
+    }
+
+    // Now, do active NPCs.
+    for( npc &guy : g->all_npcs() ) {
+        int turns = 0;
+        if( guy.is_mounted() ) {
+            guy.check_mount_is_spooked();
+        }
+        m.creature_in_field( guy );
+        if( !guy.has_effect( effect_npc_suspend ) ) {
+            guy.process_turn();
+        }
+        while( !guy.is_dead() && ( !guy.in_sleep_state() || guy.activity.id() == ACT_OPERATION ) &&
+               guy.moves > 0 && turns < 10 ) {
+            int moves = guy.moves;
+            guy.move();
+            if( moves == guy.moves ) {
+                // Count every time we exit npc::move() without spending any moves.
+                turns++;
+            }
+
+            // Turn on debug mode when in infinite loop
+            // It has to be done before the last turn, otherwise
+            // there will be no meaningful debug output.
+            if( turns == 9 ) {
+                debugmsg( "NPC %s entered infinite loop.  Turning on debug mode",
+                          guy.name );
+                debug_mode = true;
+                // make sure the filter is active
+                if( std::find(
+                        debugmode::enabled_filters.begin(), debugmode::enabled_filters.end(),
+                        debugmode::DF_NPC ) == debugmode::enabled_filters.end() ) {
+                    debugmode::enabled_filters.emplace_back( debugmode::DF_NPC );
+                }
+            }
+        }
+
+        // If we spun too long trying to decide what to do (without spending moves),
+        // Invoke cognitive suspension to prevent an infinite loop.
+        if( turns == 10 ) {
+            add_msg( _( "%s faints!" ), guy.name );
+            guy.reboot();
+        }
+
+        if( !guy.is_dead() ) {
+            guy.npc_update_body();
+        }
+    }
+    g->cleanup_dead();
+}
+
+void overmap_npc_move()
+{
+    avatar &u = get_avatar();
+    std::vector<npc *> travelling_npcs;
+    static constexpr int move_search_radius = 600;
+    for( auto &elem : overmap_buffer.get_npcs_near_player( move_search_radius ) ) {
+        if( !elem ) {
+            continue;
+        }
+        npc *npc_to_add = elem.get();
+        if( ( !npc_to_add->is_active() || rl_dist( u.pos(), npc_to_add->pos() ) > SEEX * 2 ) &&
+            npc_to_add->mission == NPC_MISSION_TRAVELLING ) {
+            travelling_npcs.push_back( npc_to_add );
+        }
+    }
+    bool npcs_need_reload = false;
+    for( auto &elem : travelling_npcs ) {
+        if( elem->has_omt_destination() ) {
+            if( !elem->omt_path.empty() ) {
+                if( rl_dist( elem->omt_path.back(), elem->global_omt_location() ) > 2 ) {
+                    // recalculate path, we got distracted doing something else probably
+                    elem->omt_path.clear();
+                } else if( elem->omt_path.back() == elem->global_omt_location() ) {
+                    elem->omt_path.pop_back();
+                }
+            }
+            if( elem->omt_path.empty() ) {
+                elem->omt_path = overmap_buffer.get_npc_path( elem->global_omt_location(), elem->goal );
+                if( elem->omt_path.empty() ) { // goal is unreachable, or already reached goal, reset it
+                    elem->goal = npc::no_goal_point;
+                }
+            } else {
+                // TODO: fix point types
+                elem->travel_overmap( project_to<coords::sm>( elem->omt_path.back() ).raw() );
+                npcs_need_reload = true;
+            }
+        }
+        if( !elem->has_omt_destination() && calendar::once_every( 1_hours ) && one_in( 3 ) ) {
+            // travelling destination is reached/not set, try different one
+            elem->set_omt_destination();
+        }
+    }
+    if( npcs_need_reload ) {
+        g->reload_npcs();
+    }
+}
+
+} // namespace
+
+// MAIN GAME LOOP
+// Returns true if game is over (death, saved, quit, etc)
+bool do_turn()
+{
+    if( g->is_game_over() ) {
+        return turn_handler::cleanup_at_end();
+    }
+    // Actual stuff
+    if( g-> new_game ) {
+        g->new_game = false;
+    } else {
+        g->gamemode->per_turn();
+        calendar::turn += 1_turns;
+    }
+
+    weather_manager &weather = get_weather();
+    // starting a new turn, clear out temperature cache
+    weather.temperature_cache.clear();
+
+    if( g->npcs_dirty ) {
+        g->load_npcs();
+    }
+
+    timed_event_manager &timed_events = get_timed_events();
+    timed_events.process();
+    mission::process_all();
+    avatar &u = get_avatar();
+    map &m = get_map();
+    // If controlling a vehicle that is owned by someone else
+    if( u.in_vehicle && u.controlling_vehicle ) {
+        vehicle *veh = veh_pointer_or_null( m.veh_at( u.pos() ) );
+        if( veh && !veh->handle_potential_theft( dynamic_cast<player &>( u ), true ) ) {
+            veh->handle_potential_theft( dynamic_cast<player &>( u ), false, false );
+        }
+    }
+    // If riding a horse - chance to spook
+    if( u.is_mounted() ) {
+        u.check_mount_is_spooked();
+    }
+    if( calendar::once_every( 1_days ) ) {
+        overmap_buffer.process_mongroups();
+    }
+
+    // Move hordes every 2.5 min
+    if( calendar::once_every( time_duration::from_minutes( 2.5 ) ) ) {
+        overmap_buffer.move_hordes();
+        // Hordes that reached the reality bubble need to spawn,
+        // make them spawn in invisible areas only.
+        m.spawn_monsters( false );
+    }
+
+    g->debug_hour_timer.print_time();
+
+    u.update_body();
+
+    // Auto-save if autosave is enabled
+    if( get_option<bool>( "AUTOSAVE" ) &&
+        calendar::once_every( 1_turns * get_option<int>( "AUTOSAVE_TURNS" ) ) &&
+        !u.is_dead_state() ) {
+        g->autosave();
+    }
+
+    weather.update_weather();
+    g->reset_light_level();
+
+    g->perhaps_add_random_npc();
+    while( u.moves > 0 && u.activity ) {
+        u.activity.do_turn( u );
+    }
+    // Process NPC sound events before they move or they hear themselves talking
+    for( npc &guy : g->all_npcs() ) {
+        if( rl_dist( guy.pos(), u.pos() ) < MAX_VIEW_DISTANCE ) {
+            sounds::process_sound_markers( &guy );
+        }
+    }
+
+    // Process sound events into sound markers for display to the player.
+    sounds::process_sound_markers( &u );
+
+    if( u.is_deaf() ) {
+        sfx::do_hearing_loss();
+    }
+
+    if( !u.has_effect( efftype_id( "sleep" ) ) || g->uquit == QUIT_WATCH ) {
+        if( u.moves > 0 || g->uquit == QUIT_WATCH ) {
+            while( u.moves > 0 || g->uquit == QUIT_WATCH ) {
+                g->cleanup_dead();
+                g->mon_info_update();
+                // Process any new sounds the player caused during their turn.
+                for( npc &guy : g->all_npcs() ) {
+                    if( rl_dist( guy.pos(), u.pos() ) < MAX_VIEW_DISTANCE ) {
+                        sounds::process_sound_markers( &guy );
+                    }
+                }
+                sounds::process_sound_markers( &u );
+                if( !u.activity && !u.has_distant_destination() && g->uquit != QUIT_WATCH ) {
+                    g->wait_popup.reset();
+                    ui_manager::redraw();
+                }
+
+                if( g->handle_action() ) {
+                    ++g->moves_since_last_save;
+                    u.action_taken();
+                }
+
+                if( g->is_game_over() ) {
+                    return turn_handler::cleanup_at_end();
+                }
+
+                if( g->uquit == QUIT_WATCH ) {
+                    break;
+                }
+                while( u.moves > 0 && u.activity ) {
+                    u.activity.do_turn( u );
+                }
+            }
+            // Reset displayed sound markers now that the turn is over.
+            // We only want this to happen if the player had a chance to examine the sounds.
+            sounds::reset_markers();
+        } else {
+            // Rate limit key polling to 10 times a second.
+            static auto start = std::chrono::time_point_cast<std::chrono::milliseconds>(
+                                    std::chrono::system_clock::now() );
+            const auto now = std::chrono::time_point_cast<std::chrono::milliseconds>(
+                                 std::chrono::system_clock::now() );
+            if( ( now - start ).count() > 100 ) {
+                handle_key_blocking_activity();
+                start = now;
+            }
+
+            g->mon_info_update();
+
+            // If player is performing a task, a monster is dangerously close,
+            // and monster can reach to the player or it has some sort of a ranged attack,
+            // warn them regardless of previous safemode warnings
+            if( u.activity && !u.has_activity( activity_id( "ACT_AIM" ) ) &&
+                u.activity.moves_left > 0 &&
+                !u.activity.is_distraction_ignored( distraction_type::hostile_spotted_near ) ) {
+                Creature *hostile_critter = g->is_hostile_very_close( true );
+
+                if( hostile_critter != nullptr ) {
+                    g->cancel_activity_or_ignore_query( distraction_type::hostile_spotted_near,
+                                                        string_format( _( "The %s is dangerously close!" ),
+                                                                hostile_critter->get_name() ) );
+                }
+            }
+
+        }
+    }
+
+    if( g->driving_view_offset.x != 0 || g->driving_view_offset.y != 0 ) {
+        // Still have a view offset, but might not be driving anymore,
+        // or the option has been deactivated,
+        // might also happen when someone dives from a moving car.
+        // or when using the handbrake.
+        vehicle *veh = veh_pointer_or_null( m.veh_at( u.pos() ) );
+        g->calc_driving_offset( veh );
+    }
+
+    scent_map &scent = get_scent();
+    // No-scent debug mutation has to be processed here or else it takes time to start working
+    if( !u.has_flag( STATIC( json_character_flag( "NO_SCENT" ) ) ) ) {
+        scent.set( u.pos(), u.scent, u.get_type_of_scent() );
+        overmap_buffer.set_scent( u.global_omt_location(),  u.scent );
+    }
+    scent.update( u.pos(), m );
+
+    // We need floor cache before checking falling 'n stuff
+    m.build_floor_caches();
+
+    m.process_falling();
+    m.vehmove();
+    m.process_fields();
+    m.process_items();
+    explosion_handler::process_explosions();
+    m.creature_in_field( u );
+
+    // Apply sounds from previous turn to monster and NPC AI.
+    sounds::process_sounds();
+    const int levz = m.get_abs_sub().z;
+    // Update vision caches for monsters. If this turns out to be expensive,
+    // consider a stripped down cache just for monsters.
+    m.build_map_cache( levz, true );
+    monmove();
+    if( calendar::once_every( 5_minutes ) ) {
+        overmap_npc_move();
+    }
+    if( calendar::once_every( 10_seconds ) ) {
+        for( const tripoint &elem : m.get_furn_field_locations() ) {
+            const furn_t &furn = *m.furn( elem );
+            for( const emit_id &e : furn.emissions ) {
+                m.emit_field( elem, e );
+            }
+        }
+        for( const tripoint &elem : m.get_ter_field_locations() ) {
+            const ter_t &ter = *m.ter( elem );
+            for( const emit_id &e : ter.emissions ) {
+                m.emit_field( elem, e );
+            }
+        }
+    }
+    turn_handler::update_stair_monsters();
+    g->mon_info_update();
+    u.process_turn();
+    if( u.moves < 0 && get_option<bool>( "FORCE_REDRAW" ) ) {
+        ui_manager::redraw();
+        refresh_display();
+    }
+    effect_on_conditions::process_effect_on_conditions();
+
+    if( levz >= 0 && !u.is_underwater() ) {
+        handle_weather_effects( weather.weather_id );
+    }
+
+    const bool player_is_sleeping = u.has_effect( effect_sleep );
+    bool wait_redraw = false;
+    std::string wait_message;
+    time_duration wait_refresh_rate;
+    if( player_is_sleeping ) {
+        wait_redraw = true;
+        wait_message = _( "Wait till you wake up…" );
+        wait_refresh_rate = 30_minutes;
+    } else if( const cata::optional<std::string> progress = u.activity.get_progress_message( u ) ) {
+        wait_redraw = true;
+        wait_message = *progress;
+        if( u.activity.is_interruptible() && u.activity.interruptable_with_kb ) {
+            wait_message += string_format( _( "\n%s to interrupt" ), press_x( ACTION_PAUSE ) );
+        }
+        wait_refresh_rate = 5_minutes;
+    }
+    if( wait_redraw ) {
+        if( g->first_redraw_since_waiting_started || calendar::once_every( 1_minutes ) ) {
+            if( g->first_redraw_since_waiting_started || calendar::once_every( wait_refresh_rate ) ) {
+                ui_manager::redraw();
+            }
+
+            // Avoid redrawing the main UI every time due to invalidation
+            ui_adaptor dummy( ui_adaptor::disable_uis_below {} );
+            g->wait_popup = std::make_unique<static_popup>();
+            g->wait_popup->on_top( true ).wait_message( "%s", wait_message );
+            ui_manager::redraw();
+            refresh_display();
+            g->first_redraw_since_waiting_started = false;
+        }
+    } else {
+        // Nothing to wait for now
+        g->wait_popup.reset();
+        g->first_redraw_since_waiting_started = true;
+    }
+
+    u.update_bodytemp();
+    u.update_body_wetness( *weather.weather_precise );
+    u.apply_wetness_morale( weather.temperature );
+
+    if( calendar::once_every( 1_minutes ) ) {
+        u.update_morale();
+    }
+
+    if( calendar::once_every( 9_turns ) ) {
+        u.check_and_recover_morale();
+    }
+
+    if( !u.is_deaf() ) {
+        sfx::remove_hearing_loss();
+    }
+    sfx::do_danger_music();
+    sfx::do_vehicle_engine_sfx();
+    sfx::do_vehicle_exterior_engine_sfx();
+    sfx::do_fatigue();
+
+    // reset player noise
+    u.volume = 0;
+
+    return false;
+}
