+++ src/do_turn.h
@@ -1,0 +1,8 @@
+#pragma once
+#ifndef CATA_SRC_DO_TURN_H
+#define CATA_SRC_DO_TURN_H
+
+/** MAIN GAME LOOP. Returns true if game is over (death, saved, quit, etc.). */
+bool do_turn();
+
+#endif // CATA_SRC_DO_TURN_H
